refactor(api): extract shared retry/error constants to danyapi/retry.py
Move MAX_RETRIES, RETRY_BACKOFF_SEC, RETRY_BACKOFF_MAX_SEC,
RETRYABLE_HTTP_STATUSES, _is_retryable_http and _drop_session from
danyapi/api/deepseek.py and danyapi/qwen/api.py into a single shared
module. Both consumers import them as local names so existing
monkeypatch-based tests (openai_mod.RETRY_BACKOFF_SEC / qwen_api.*)
continue to work.
